# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing database connections and syncing data to destinations like Typesense. List endpoints for sync jobs and sync logs return a paginated envelope: items, page, page_size, total, total_pages.",
        "title": "Portico Sync API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/sync-jobs": {
            "get": {
                "description": "Returns a paginated list of sync jobs. Default page size is 20; maximum is 100.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync-jobs"
                ],
                "summary": "List sync jobs",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page (max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SyncJobListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync-jobs"
                ],
                "summary": "Create sync job",
                "parameters": [
                    {
                        "description": "Sync job",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/syncjob.CreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SyncJob"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sync-jobs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync-jobs"
                ],
                "summary": "Get sync job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SyncJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync-jobs"
                ],
                "summary": "Update sync job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sync job",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/syncjob.UpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SyncJob"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "sync-jobs"
                ],
                "summary": "Delete sync job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sync-jobs/{id}/explore": {
            "post": {
                "description": "Returns a paginated preview of source (rules/fields/relations applied) or destination documents for a sync job.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync-jobs"
                ],
                "summary": "Explore sync job data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Explore options",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ExploreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sync.PreviewResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sync-jobs/{id}/explore/export": {
            "post": {
                "description": "Downloads matching source or destination rows as CSV.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "sync-jobs"
                ],
                "summary": "Export sync job explore data as CSV",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Export options",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ExploreExportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sync-jobs/{id}/run": {
            "post": {
                "description": "Runs the sync synchronously and returns the finished sync log.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync-jobs"
                ],
                "summary": "Run a sync job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SyncLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sync-jobs/{id}/start": {
            "post": {
                "description": "Creates a sync log with status running and returns it immediately. Poll GET /sync-logs/{id} for progress (rows_synced / rows_total) and final status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync-jobs"
                ],
                "summary": "Start a sync job in the background",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.SyncLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.ExploreExportRequest": {
            "type": "object",
            "required": [
                "side"
            ],
            "properties": {
                "side": {
                    "type": "string",
                    "example": "source"
                }
            }
        },
        "handlers.ExploreRequest": {
            "type": "object",
            "required": [
                "side"
            ],
            "properties": {
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 50
                },
                "side": {
                    "type": "string",
                    "example": "source"
                },
                "sort_by": {
                    "type": "string",
                    "example": "id"
                },
                "sort_dir": {
                    "type": "string",
                    "example": "asc"
                }
            }
        },
        "handlers.SyncJobListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SyncJob"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 42
                },
                "total_pages": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "models.Connection": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.SyncJob": {
            "type": "object",
            "properties": {
                "chunk_size": {
                    "type": "integer"
                },
                "config": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "destination_connection": {
                    "$ref": "#/definitions/models.Connection"
                },
                "destination_connection_id": {
                    "type": "integer"
                },
                "destination_table": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SyncJobField"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "relations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SyncJobRelation"
                    }
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SyncJobRule"
                    }
                },
                "source_connection": {
                    "$ref": "#/definitions/models.Connection"
                },
                "source_connection_id": {
                    "type": "integer"
                },
                "source_table": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workers": {
                    "type": "integer"
                }
            }
        },
...
        "models.SyncJobRule": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "operator": {
                    "type": "string"
                },
                "sync_job_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.SyncLog": {
            "type": "object",
            "properties": {
                "duration_ms": {
                    "type": "integer"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "rows_synced": {
                    "type": "integer"
                },
                "rows_total": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "sync_job": {
                    "$ref": "#/definitions/models.SyncJob"
                },
                "sync_job_id": {
                    "type": "integer"
                }
            }
        },
        "sync.PreviewResult": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                },
                "side": {
                    "type": "string"
                },
                "sort_by": {
                    "type": "string"
                },
                "sort_dir": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "syncjob.CreateInput": {
            "type": "object",
            "required": [
                "destination_connection_id",
                "destination_table",
                "name",
                "source_connection_id",
                "source_table"
            ],
            "properties": {
                "chunk_size": {
                    "type": "integer"
                },
                "config": {
                    "type": "object"
                },
                "destination_connection_id": {
                    "type": "integer"
                },
                "destination_table": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/syncjob.FieldInput"
                    }
                },
                "name": {
                    "type": "string"
                },
                "relations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/syncjob.RelationInput"
                    }
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/syncjob.RuleInput"
                    }
                },
                "source_connection_id": {
                    "type": "integer"
                },
                "source_table": {
                    "type": "string"
                },
                "workers": {
                    "type": "integer"
                }
            }
        },
        "syncjob.FieldInput": {
            "type": "object",
            "required": [
                "source_name"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "destination_config": {
                    "type": "object"
                },
                "destination_name": {
                    "type": "string"
                },
                "destination_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "source_name": {
                    "type": "string"
                },
                "sync_job_relation_id": {
                    "type": "integer"
                }
            }
        },
        "syncjob.RelationInput": {
            "type": "object",
            "required": [
                "name",
                "table",
                "type"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "config": {
                    "type": "object"
                },
                "foreign_key": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "related_key": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "syncjob.RuleInput": {
            "type": "object",
            "required": [
                "field",
                "operator"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "field": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "operator": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "syncjob.UpdateInput": {
            "type": "object",
            "properties": {
                "chunk_size": {
                    "type": "integer"
                },
                "config": {
                    "type": "object"
                },
                "destination_connection_id": {
                    "type": "integer"
                },
                "destination_table": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/syncjob.FieldInput"
                    }
                },
                "name": {
                    "type": "string"
                },
                "relations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/syncjob.RelationInput"
                    }
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/syncjob.RuleInput"
                    }
                },
                "source_connection_id": {
                    "type": "integer"
                },
                "source_table": {
                    "type": "string"
                },
                "workers": {
                    "type": "integer"
                }
            }
        }
    }
}